Cleaned up and added docstrings
A lot of the code has been cleaned up and refactored. Doc Strings have
been added to many classes. item.py and event.py haven't been documented
yet, that'll be done when those classes are fully implemented. Likewise
with display.py, when the code for n dimensional displaying is set up
the code will be refactored and documented at the same time.